requirements.txt
click==8.0.3
colorama==0.4.4
cycler==0.11.0
Flask==2.0.2
Flask-WTF==1.0.0
fonttools==4.28.5
itsdangerous==2.0.1
Jinja2==3.0.3
kiwisolver==1.3.2
MarkupSafe==2.0.1
matplotlib==3.5.1
numpy==1.21.5
packaging==21.3
Pillow==8.4.0
pyparsing==3.0.6
python-dateutil==2.8.2
python-dotenv==0.19.2
six==1.16.0
Werkzeug==2.0.2
WTForms==3.0.1
